feat: improve event and delegate colors in themes
This commit improves the color representation of events and delegates across all Calmppuccin themes. The event color has been changed to a more distinct green (`#8de652` in Latte, `#a8dba1` in Mocha, `#a8e0a0` in Frappe, and `#a8dda1` in Macchiato) for better visibility. The delegate color has been adjusted to a slightly different shade of green (`#52e67e` in Latte, `#a1dbc0` in Mocha, `#a4e6c7` in Frappe, and `#a4e0c4` in Macchiato) to visually differentiate it from events. These changes enhance the overall readability and aesthetics of the themes, making it easier to identify and distinguish events and delegates in the code.